Dodana funkcionalna stranica za login i dodana restrikcija korisnika na odeđene stranice

+++ gd_forum/base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
 from .forms import NewUserForm
+from django.http import HttpResponse
 from django.contrib.auth.forms import UserCreationForm 
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import login, authenticate, logout
 from django.contrib import messages
 from .models import Thread, Category
@@ -18,6 +20,8 @@ def register_user (request):
             login(request, user)
             messages.success(request, 'Registration Successful!')
             return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
     else:
         form = UserCreationForm()
     return render(request, 'base/register.html', context={'form': form})
@@ -27,7 +31,26 @@ def logout_user(request):
     return redirect('home')
     
 def login_user(request):
-    pass
+    if request.user.is_authenticated:
+        return redirect( 'home' )
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+        if (User is not None):
+            login(request, user)
+            return redirect( 'home' )
+
+        else:
+            return messages.error(request, 'Username or password are incorrect')
+    context = {}
+    return render(request, 'base/login.html', context)
 
 def home (request):
     threads = Thread.objects.all()
@@ -55,6 +78,7 @@ def user_view(request, name):
     }
     return render(request, 'base/user_view.html', context)
 
+@login_required(login_url='login')
 def create_thread(request):
     form = ThreadForm()
     if request.method == 'POST':
@@ -68,8 +92,12 @@ def create_thread(request):
     }
     return render(request, 'base/thread_form.html', context)
 
+@login_required(login_url='login')
 def delete_thread(request, pk):
     thread = Thread.objects.get(id=pk)
+    if request.user != thread.uploader:
+        return redirect ('home')
+
     if request.method == "POST":
         thread.delete()
         return redirect ( 'home' )
